Move ut_linci into confidence, and make _confidence private.
