+++ utide/confidence.py
@@ -1,14 +1,31 @@
+"""
+Confidence interval calculations for solve().
+
+Includes a strictly private _confidence() function and
+a more general ut_linci() function for linearized estimates
+of ellipse parameter uncertainties.
+"""
+
 from __future__ import absolute_import, division
 
 import numpy as np
 import scipy.interpolate as sip
 
 from .ut_pdgm import ut_pdgm
-from .ut_linci import ut_linci
 
 
-def ut_confidence(coef, opt, t, e, tin, tgd, uvgd, elor, xraw, xmod, W, m, B,
+def _confidence(coef, opt, t, e, tin, tgd, uvgd, elor, xraw, xmod, W, m, B,
                   nm, nt, nc, Xu, Yu, Xv, Yv):
+    """
+    This confidence interval calculation does not correspond
+    to a single ut_ matlab function, but is based on code
+    in ut_solv1 starting from the line
+
+    '%% spectral power (Puu, Pvv, Puv) of residual'.
+
+    It returns its first argument, the coef dictionary,
+    with additional entries.
+    """
 
     # Confidence Intervals.
     print('conf. int vls... ')
@@ -165,3 +182,97 @@ def ut_confidence(coef, opt, t, e, tin, tgd, uvgd, elor, xraw, xmod, W, m, B,
             pass
 
     return coef
+
+
+def ut_linci(X, Y, sigX, sigY):
+    # UT_LINCI()
+    # current ellipse parameter uncertainties from cosine/sine coefficient
+    # uncertainties, by linearized relations w/ correlations presumed zero
+    # inputs: (two-dim case complex, one-dim case real)
+    # X = Xu + i*Xv
+    # Y = Yu + i*Yv
+    # for Xu =real(X) = u cosine coeff; Yu =real(Y) = u sine coeff
+    # Xv =imag(X) = v cosine coeff; Yv =imag(Y) = v sine coeff
+    # sigX = sigXu + i*sigXv
+    # sigY = sigYu + i*sigYv
+    # for sigXu =real(sigX) =stddev(Xu); sigYu =real(sigY) =stddev(Yu)
+    # sigXv =imag(sigX) =stddev(Xv); sigYv =imag(sigY) =stddev(Yv)
+    # outputs:
+    # two-dim case, complex
+    # sig1 = sig_Lsmaj +1i*sig_Lsmin [same units as inputs]
+    # sig2 = sig_g + 1i*sig_theta [degrees]
+    # one-dim case, real
+    # sig1 = sig_A [same units as inputs]
+    # sig2 = sig_g [degrees]
+    # UTide v1p0 9/2011 d.codiga@gso.uri.edu
+    # (adapted from errell.m of t_tide, Pawlowicz et al 2002)
+
+    X = np.array([X])
+    Y = np.array([Y])
+    sigX = np.array([sigX])
+    sigY = np.array([sigY])
+    Xu = np.real(X[:])
+    sigXu = np.real(sigX)
+    Yu = np.real(Y[:])
+    sigYu = np.real(sigY)
+
+    Xv = np.imag(X[:])
+    sigXv = np.imag(sigX[:])
+    Yv = np.imag(Y[:])
+    sigYv = np.imag(sigY[:])
+
+    rp = 0.5 * np.sqrt((Xu+Yv)**2 + (Xv-Yu)**2)
+    rm = 0.5 * np.sqrt((Xu-Yv)**2 + (Xv+Yu)**2)
+    sigXu2 = sigXu**2
+    sigYu2 = sigYu**2
+    sigXv2 = sigXv**2
+    sigYv2 = sigYv**2
+
+    ex = (Xu+Yv) / rp
+    fx = (Xu-Yv) / rm
+    gx = (Yu-Xv) / rp
+    hx = (Yu+Xv) / rm
+
+    # major axis
+    dXu2 = (0.25*(ex+fx))**2
+    dYu2 = (0.25*(gx+hx))**2
+    dXv2 = (0.25*(hx-gx))**2
+    dYv2 = (0.25*(ex-fx))**2
+    sig1 = np.sqrt(dXu2 * sigXu2 + dYu2 * sigYu2 +
+                   dXv2 * sigXv2 + dYv2 * sigYv2)
+
+    # phase
+    rn = 2 * (Xu * Yu + Xv * Yv)
+    rd = Xu**2 - Yu**2 + Xv**2 - Yv**2
+    den = rn**2 + rd**2
+    dXu2 = ((rd*Yu - rn*Xu) / den)**2
+    dYu2 = ((rd*Xu + rn*Yu) / den)**2
+    dXv2 = ((rd*Yv - rn*Xv) / den)**2
+    dYv2 = ((rd*Xv + rn*Yv) / den)**2
+    sig2 = (180/np.pi) * np.sqrt(dXu2 * sigXu2 + dYu2 * sigYu2 +
+                                 dXv2 * sigXv2 + dYv2 * sigYv2)
+
+    # if ~isreal(X)
+    if not np.isreal(X):
+        # Minor axis.
+        dXu2 = (0.25 * (ex-fx))**2
+        dYu2 = (0.25 * (gx-hx))**2
+        dXv2 = (0.25 * (hx+gx))**2
+        dYv2 = (0.25 * (ex+fx))**2
+        sig1 = sig1 + 1j*np.sqrt(dXu2 * sigXu2 + dYu2 * sigYu2 +
+                                 dXv2 * sigXv2 + dYv2 * sigYv2)
+
+        # Orientation.
+        rn = 2.0 * (Xu * Xv + Yu * Yv)
+        rd = Xu**2 + Yu**2 - (Xv**2 + Yv**2)
+        den = rn**2 + rd**2
+        dXu2 = ((rd*Xv - rn*Xu) / den)**2
+        dYu2 = ((rd*Yv - rn*Yu) / den)**2
+        dXv2 = ((rd*Xu + rn*Xv) / den)**2
+        dYv2 = ((rd*Yu + rn*Yv) / den)**2
+        sig2 = sig2 + 1j*(180/np.pi) * np.sqrt(dXu2 * sigXu2 + dYu2 * sigYu2 +
+                                               dXv2 * sigXv2 + dYv2 * sigYv2)
+
+    return sig1, sig2
+
+
